+++ libgraphal/valuebool.h
@@ -56,7 +56,7 @@ public:
 	virtual bool     toBool(void) const { return m_val; }
 	virtual string toString(void) const { return (m_val) ? "true" : "false"; }
 	virtual bool  isNumeric(void) const { return true; }
-	virtual float     toInt(void) const { return m_val; }
+	virtual int       toInt(void) const { return m_val; }
 	virtual float   toFloat(void) const { return m_val; }
 
 	virtual ValueBool*           toValueBool(void) { return this; }
